requirements.txt
amqp==5.3.1
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
billiard==4.2.1
celery==5.5.3
certifi==2025.6.15
cffi==1.17.1
channels==4.2.2
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==45.0.5
daphne==4.2.1
dj-rest-auth==7.0.1
Django==5.2.3
django-allauth==65.9.0
django-celery-beat==2.8.1
django-filter==25.1
django-redis==6.0.0
django-timezone-field==7.1
django-unfold==0.60.0
djangorestframework==3.16.0
geographiclib==2.0
geopy==2.4.1
h11==0.16.0
httptools==0.6.4
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
kombu==5.5.4
packaging==25.0
pillow==11.2.1
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pyOpenSSL==25.1.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
redis==6.2.0
requests==2.32.4
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
vine==5.1.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
zope.interface==7.2
